pyproject.toml
[tool.poetry]
name = "mssbot"
version = "0.1.0"
description = ""
authors = ["JoJoJux <80523427+J0J0HA@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nextcord = "^2.6.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
